restructure: put all files in one folder
